custom_os/requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
Jinja2==3.1.3
MarkupSafe==2.1.5
pyelftools==0.31
PyYAML==6.0.1
requests==2.31.0
termcolor==2.4.0
urllib3==2.2.1
pyyaml==6.0.1
